Add new endpoint to update note records, disallow GET, PUT for specific note records
To be consistent, I am using PUT throughout this API to replace objects, so PUT is disallowed already for specific notes. The endpoint to retrieve a specific note would be /assets/:asset_id/notes/:note_id, it doesn't make sense to me to retrieve a single especically if you already know the asset's id; therefore I am leaving this out.